Renamed some stuff, added shaders to new gl package

+++ src/lb/simplebase/gl/GLHandle.java
@@ -2,6 +2,6 @@ package lb.simplebase.gl;
 
 public interface GLHandle {
 
-	public long getGLHandle();
+	public int getGLHandle();
 	
 }
